util/av: Handle NULL address insertion
An MPI test failed because it called fi_av_insert() with a count
of 0 and an invalid address.  Because the address was invalid, the
code assigned it an address length of 0.  This in turn resulted in
the AV setting its internal addrlen to 0, since 0 was smaller than
the current address.

To fix this issue and make the code more robust, define a new
internal AV flag.  When set, it indicates that the addrlen needs
to be determined based on the first address insertion.  After
setting the addrlen, clear the flag to prevent further updates.
As part of the change, we enhance the checks and fail addresses
that are an incorrect length.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>