prov/shm: Combine tx_entry and sar_entry into progress entry
Generify these entries for future asynchronous applications.
They are very similar and are both used for pending operations.
They can be the same for send and receive.

Rename peer_id to send_id to make it clear that its only applicable
for the send side.

Merged tx and sar entry struct can simplify functions in DSA and use
the send_id to identify how to use it.

Add an init_rx_pend function to simplify receive progress entry
initialization which initializes send_id = -1 to indicate a receive.

Signed-off-by: zdworkin <zachary.dworkin@intel.com>