contrib/intel/jenkins: Update check_fail case for daos in summarizer
Changing the check_fail case for daos to check for fail.
Currently if there is a failure it is reported on:
error, fail, skip, warn, interrupt, and cancel.

This causes the summary to print out the failing test's name 5 times instead of one.
Updating this to check for fail and to check for not pass and not fail will fix this.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>