prov/shm: protect map ze pid_fd lookups with id
Coverity was complaining that set_ipc_valid needed to be locked
because it looks up a value in the region->map that can change
after being checked.

This id is now set in map_to_region instead of map_add.
The id is only going to be valid if the region is successfully
mapped.
map_to_region can be called in 3 cases: map_add, verify_peer,
and progress_connreq's map_add.
In all of these cases the set_ipc_valid value in the map lookup
is always valid because it will never be called with an
invalid id.

There will need to be future work done to protect against an
edge case discovered while making these changes.
This edge case is if a peer destroyed their region and then
tried to re-open one on with the same pid and same name. This
will result in a peer that has a valid id but an invalid
region. The Libfabric API describes this as undefined behaviour
but we can solve this edge case by adding an unmap_region
function in the future.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>