prov/shm: Add unmap_peer function
unmap_peer is useful for when a thread needs to cleanup a peer
and then re-open a new one with the same name. This function
fixes the current bug where if a thread want's to cleanup a
peer's region it unmaps it without cleaning up any other
resources leading to a valid ID with an invalid/nonexistant
shm region.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>