contrib/intel/jenkins: Send logs at the end of a stage
Send logs at the end of a stage instead of gathering
them in the summary stage. This is possible only after
migrating stages to use RED. It allows logs to be on
the main node even if a stage failed so that the post
summary stage does not have to duplicate gathering logs
if everything passed.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>