include/atomic: Push claims back on discard path
Move the push of available claims from the claim
path to the discard path. This makes it so that
the discard variable doesn't need to be an atomic,
and makes it so that the receiver is in charge of
putting the claims back so the sender can keep
its normal flow of transmitting things.

Introduce the environment variable
"SMR_ATOMIC_DISCARD_THRESHOLD" to control when the
receiver should put available claims back.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>